+++ src/main/java/org/folio/linked/data/controller/ResourceController.java
@@ -6,11 +6,14 @@ import static org.springframework.http.HttpStatus.CREATED;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
+import java.util.Set;
 import lombok.RequiredArgsConstructor;
 import org.folio.linked.data.domain.dto.ResourceIdDto;
 import org.folio.linked.data.domain.dto.ResourceMarcViewDto;
 import org.folio.linked.data.domain.dto.ResourceRequestDto;
 import org.folio.linked.data.domain.dto.ResourceResponseDto;
+import org.folio.linked.data.domain.dto.ResourceSubgraphViewDto;
+import org.folio.linked.data.domain.dto.SearchResourcesRequestDto;
 import org.folio.linked.data.rest.resource.ResourceApi;
 import org.folio.linked.data.service.rdf.RdfExportService;
 import org.folio.linked.data.service.resource.ResourceService;
@@ -79,6 +82,11 @@ public class ResourceController implements ResourceApi {
   @Override
   public ResponseEntity<String> exportInstanceToRdf(Long id) {
     return ResponseEntity.ok(rdfExportService.exportInstanceToRdf(id));
+  }
+
+  @Override
+  public ResponseEntity<Set<ResourceSubgraphViewDto>> searchResources(SearchResourcesRequestDto searchRequest) {
+    return ResponseEntity.ok(resourceService.searchResources(searchRequest));
   }
 
   @InitBinder