+++ src/main/java/org/folio/linked/data/service/resource/ResourceService.java
@@ -4,6 +4,8 @@ import java.util.Set;
 import org.folio.linked.data.domain.dto.ResourceIdDto;
 import org.folio.linked.data.domain.dto.ResourceRequestDto;
 import org.folio.linked.data.domain.dto.ResourceResponseDto;
+import org.folio.linked.data.domain.dto.ResourceSubgraphViewDto;
+import org.folio.linked.data.domain.dto.SearchResourcesRequestDto;
 
 public interface ResourceService {
 
@@ -12,6 +14,8 @@ public interface ResourceService {
   ResourceResponseDto getResourceById(Long id);
 
   ResourceIdDto getResourceIdByInventoryId(String inventoryId);
+
+  Set<ResourceSubgraphViewDto> searchResources(SearchResourcesRequestDto searchRequest);
 
   ResourceResponseDto updateResource(Long id, ResourceRequestDto bibframeRequest);